requirements.txt
ackermann-msgs==2.0.2
action-msgs==1.2.2
action-tutorials-interfaces==0.20.8
action-tutorials-py==0.20.8
actionlib-msgs==4.9.1
altair==6.0.0
ament-cmake-test==1.3.13
ament-copyright==0.12.15
ament-cppcheck==0.12.15
ament-cpplint==0.12.15
ament-flake8==0.12.15
ament-index-python==1.4.1
ament-lint==0.12.15
ament-lint-cmake==0.12.15
ament-package==0.14.2
ament-pep257==0.12.15
ament-uncrustify==0.12.15
ament-xmllint==0.12.15
angles==1.15.0
attrs==25.4.0
blinker==1.9.0
bond==4.1.2
builtin-interfaces==1.2.2
cachetools==7.0.5
camera-calibration==3.0.9
cartographer-ros-msgs==2.0.9002
certifi==2026.2.25
charset-normalizer==3.4.5
click==8.3.1
composition-interfaces==1.2.2
contourpy==1.3.2
control-msgs==4.8.0
controller-manager==2.53.1
controller-manager-msgs==2.53.1
cv-bridge==3.2.1
cycler==0.12.1
demo-nodes-py==0.20.8
diagnostic-msgs==4.9.1
diagnostic-updater==4.0.6
domain-coordinator==0.10.0
dwb-msgs==1.1.20
dynamixel-easy-sdk==4.0.3
dynamixel-sdk==4.0.3
example-interfaces==0.9.3
examples-rclpy-executors==0.15.5
examples-rclpy-minimal-action-client==0.15.5
examples-rclpy-minimal-action-server==0.15.5
examples-rclpy-minimal-client==0.15.5
examples-rclpy-minimal-publisher==0.15.5
examples-rclpy-minimal-service==0.15.5
examples-rclpy-minimal-subscriber==0.15.5
fonttools==4.61.1
gazebo-msgs==3.9.0
generate-parameter-library-py==0.6.0
geometry-msgs==4.9.1
gitdb==4.0.12
GitPython==3.1.46
idna==3.11
image-geometry==3.2.1
interactive-markers==2.3.2
Jinja2==3.1.6
joint-state-publisher==2.4.0
joint-state-publisher-gui==2.4.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
laser-geometry==2.4.1
launch==1.0.14
launch-ros==0.19.13
launch-testing==1.0.14
launch-testing-ros==0.19.13
launch-xml==1.0.14
launch-yaml==1.0.14
lifecycle-msgs==1.2.2
logging-demo==0.20.8
map-msgs==2.1.0
MarkupSafe==3.0.3
matplotlib==3.10.8
message-filters==4.3.14
micro-ros-msgs==1.0.0
moveit-msgs==2.2.1
mysql-connector-python==9.6.0
narwhals==2.18.0
nav-2d-msgs==1.1.20
nav-msgs==4.9.1
nav2-common==1.1.20
nav2-msgs==1.1.20
nav2-simple-commander==1.0.0
numpy==2.2.6
object-recognition-msgs==2.0.0
octomap-msgs==2.0.1
osrf-pycommon==2.1.6
packaging==26.0
pandas==2.3.3
pcl-msgs==1.0.0
pendulum-msgs==0.20.8
pillow==12.1.1
plotly==6.6.0
protobuf==6.33.5
pyarrow==23.0.1
pydeck==0.9.1
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-qt-binding==1.1.3
pytz==2026.1.post1
qt-dotgraph==2.2.4
qt-gui==2.2.4
qt-gui-cpp==2.2.4
qt-gui-py-common==2.2.4
quality-of-service-demo-py==0.20.8
rcl-interfaces==1.2.2
rclpy==3.3.20
rcutils==5.1.8
referencing==0.37.0
requests==2.32.5
resource-retriever==3.1.3
rmw-dds-common==1.6.0
ros2action==0.18.17
ros2bag==0.15.16
ros2cli==0.18.17
ros2component==0.18.17
ros2controlcli==2.53.1
ros2doctor==0.18.17
ros2interface==0.18.17
ros2launch==0.19.13
ros2lifecycle==0.18.17
ros2multicast==0.18.17
ros2node==0.18.17
ros2param==0.18.17
ros2pkg==0.18.17
ros2plugin==5.1.3
ros2run==0.18.17
ros2service==0.18.17
ros2topic==0.18.17
rosbag2-interfaces==0.15.16
rosbag2-py==0.15.16
rosgraph-msgs==1.2.2
rosidl-adapter==3.1.8
rosidl-cli==3.1.8
rosidl-cmake==3.1.8
rosidl-generator-c==3.1.8
rosidl-generator-cpp==3.1.8
rosidl-generator-py==0.14.6
rosidl-parser==3.1.8
rosidl-runtime-py==0.9.3
rosidl-typesupport-c==2.0.2
rosidl-typesupport-cpp==2.0.2
rosidl-typesupport-fastrtps-c==2.2.4
rosidl-typesupport-fastrtps-cpp==2.2.4
rosidl-typesupport-introspection-c==3.1.8
rosidl-typesupport-introspection-cpp==3.1.8
rpds-py==0.30.0
rpyutils==0.2.2
rqt-action==2.0.1
rqt-bag==1.1.6
rqt-bag-plugins==1.1.6
rqt-console==2.0.3
rqt-graph==1.3.2
rqt-gui==1.1.9
rqt-gui-py==1.1.9
rqt-msg==1.2.0
rqt-plot==1.1.5
rqt-publisher==1.5.0
rqt-py-common==1.1.9
rqt-py-console==1.0.2
rqt-reconfigure==1.1.3
rqt-service-caller==1.0.5
rqt-shell==1.0.2
rqt-srv==1.0.3
rqt-topic==1.5.1
scripts==3.9.0
sensor-msgs==4.9.1
sensor-msgs-py==4.9.1
shape-msgs==4.9.1
six==1.17.0
slam-toolbox==2.6.10
smclib==4.1.2
smmap==5.0.3
srdfdom==2.0.7
sros2==0.10.9
statistics-msgs==1.2.2
std-msgs==4.9.1
std-srvs==4.9.1
stereo-msgs==4.9.1
streamlit==1.55.0
teleop-twist-keyboard==2.4.1
tenacity==9.1.4
tf-transformations==1.1.1
tf2-geometry-msgs==0.25.19
tf2-kdl==0.25.19
tf2-msgs==0.25.19
tf2-py==0.25.19
tf2-ros-py==0.25.19
tf2-sensor-msgs==0.25.19
tf2-tools==0.25.19
theora-image-transport==2.5.4
toml==0.10.2
topic-monitor==0.20.8
tornado==6.5.5
trajectory-msgs==4.9.1
turtlebot3-applications-msgs==1.0.1
turtlebot3-aruco-tracker==1.3.3
turtlebot3-automatic-parking==1.3.3
turtlebot3-automatic-parking-vision==1.3.3
turtlebot3-autorace-camera==1.2.2
turtlebot3-autorace-detect==1.2.2
turtlebot3-autorace-mission==1.2.2
turtlebot3-example==2.3.6
turtlebot3-home-service-challenge-aruco==1.0.5
turtlebot3-home-service-challenge-core==1.0.5
turtlebot3-msgs==2.4.0
turtlebot3-teleop==2.3.6
turtlebot3-yolo-object-detection==1.3.3
turtlesim==1.4.3
typing_extensions==4.15.0
tzdata==2025.3
unique-identifier-msgs==2.2.1
urdfdom-py==1.2.1
urllib3==2.6.3
visualization-msgs==4.9.1
watchdog==6.0.0
xacro==2.1.1
